Move from manually specifying rows with errors to doing so heuristically
We use the following algorithm:

1. Consider a date (onset or death) plausible if it's between 1st Jan
2020 and whatever today's date is.
2. If a date is implausible, try swapping the day and month and see if
that is now plausible.
3. If that is now plausible (and it wasn't previously), accept the
swap.
4. Output as implausible anything that the fix didn't help with and
rows with onset date after death date (if both exist).
5. Output the rows we have swapped (in case the heuristic goes wonky).

This gives plausible results for everything in the latest (29th Mar)
deaths data except finalids 9424 and 30609 (with death dates 2018-08-19
and 2019-09-18 respectively).
Note that, with the latest data, we only swap onset (not death) dates.